Refactor: Move file directories
