+++ feature/main/src/main/java/kr/co/main/calendar/ui/calendartab/diarytab/DiaryTab.kt
@@ -1,0 +1,13 @@
+package kr.co.main.calendar.ui.calendartab.diarytab
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.hilt.navigation.compose.hiltViewModel
+
+@Composable
+internal fun DiaryTabScreen(
+    modifer: Modifier = Modifier,
+    viewModel: DiaryTabViewModel = hiltViewModel(),
+) {
+
+}
