+++ feature/main/src/main/java/kr/co/main/calendar/ui/calendartab/scheduletab/ScheduleTab.kt
@@ -1,0 +1,12 @@
+package kr.co.main.calendar.ui.calendartab.scheduletab
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.hilt.navigation.compose.hiltViewModel
+
+@Composable
+internal fun ScheduleTabScreen(
+    modifier: Modifier = Modifier,
+    viewModel: ScheduleTabViewModel = hiltViewModel()
+) {
+}
